TriggerBuildFactory::hasToken
Summary:
If the application using the factory wish to know if a token exists
for such image, it's probably easier to check the current state of
the tokens array than the source configuration.

This is the case for the Notifications center, where there is a JSON
configuration file to parse to get the tokens.

We so provide a TriggerBuildFactory::hasToken boolean method to check
if our $tokens array does have the specified image as key.

Ref T959.

Test Plan: New unit tests.

Reviewers: dereckson

Maniphest Tasks: T959

Differential Revision: https://devcentral.nasqueron.org/D653